adding final project

+++ src/app/todo-item/todo-item.component.ts
@@ -14,9 +14,14 @@ export class TodoItemComponent implements OnInit {
 	@Input()
 	private todo: Todo;
 
+  public editing = false;
+
   private isReadOnly: boolean = true;
 
   public allTodos;
+
+  private todoText;
+  private todoDesc;
 
 
   constructor(private todoService: TodoService, private storageService: StorageService){
@@ -24,23 +29,27 @@ export class TodoItemComponent implements OnInit {
   };
 
   ngOnInit() {
+
+  }
+
+  private edit(e){
+    this.editing = !this.editing;
+    this.todoText = this.todo.text;
+    this.todoDesc = this.todo.description;
   }
 
   private toggleCompleted(){
     this.todoService.isCompleted(this.todo.id, this.todo.completed);
   }
 
-  toggleEdit(){
-    this.isReadOnly = !this.isReadOnly;
-  }
-
   private updateTodo(){
-    this.todoService.update(this.todo.id, this.todo.text);
-    this.toggleEdit();
+    this.todoService.update(this.todo.id, this.todoText, this.todoDesc);
   }
 
-  private removeTodo(){
-  	this.todoService.removeTodo(this.todo.id); 
+  private removeTodo(name: string){
+  	if(confirm("Are you sure you want to delete?")){
+      this.todoService.removeTodo(this.todo.id);
+    } 
   }
 
 }
